dagu/scripts/sync-ghostfolio/pyproject.toml
[project]
name = "sync-ghostfolio"
version = "0.1.0"
description = "Syncs different finance platforms with Ghostfolio"
readme = "README.md"
authors = [{ name = "Gontzal Monasterio", email = "go.monasterio@gmail.com" }]
requires-python = ">=3.14"
dependencies = [
    "dotenv>=0.9.9",
    "ghostfolio>=0.8.0",
    "httpx>=0.28.1",
    "tradernet-sdk>=2.0.0",
]

[project.scripts]
sync-ghostfolio = "sync_ghostfolio:main"

[build-system]
requires = ["uv_build>=0.9.24,<0.10.0"]
build-backend = "uv_build"
